rename scan-results file
