+++ src/electron/common/application-properties-provider.ts
@@ -3,7 +3,7 @@
 import { AppDataAdapter } from 'common/browser-adapters/app-data-adapter';
 import { ToolData } from 'common/types/store-data/unified-data-interface';
 import { androidAppTitle } from 'content/strings/application';
-import { AndroidScanResults } from 'electron/platform/android/scan-results';
+import { AndroidScanResults } from 'electron/platform/android/android-scan-results';
 
 export type ToolDataDelegate = (scanResults: AndroidScanResults) => ToolData;
