+++ src/electron/platform/android/android-scan-results.ts
