+++ src/electron/platform/android/scan-results-to-platform-data.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { PlatformData } from 'common/types/store-data/unified-data-interface';
-import { AndroidScanResults } from './scan-results';
+import { AndroidScanResults } from './android-scan-results';
 
 export type ConvertScanResultsToPlatformDataDelegate = (
     scanResults: AndroidScanResults,
