+++ src/electron/platform/android/scan-results-to-unified-rules.ts
@@ -5,7 +5,7 @@ import { UnifiedRule } from 'common/types/store-data/unified-data-interface';
 import { UUIDGenerator } from 'common/uid-generator';
 import { RuleInformation } from './rule-information';
 import { RuleInformationProviderType } from './rule-information-provider-type';
-import { AndroidScanResults } from './scan-results';
+import { AndroidScanResults } from './android-scan-results';
 
 export type ConvertScanResultsToUnifiedRulesDelegate = (
     scanResults: AndroidScanResults,
