+++ src/electron/platform/android/unified-result-builder.ts
@@ -5,7 +5,7 @@ import { generateUID, UUIDGenerator } from 'common/uid-generator';
 import { ToolDataDelegate } from 'electron/common/application-properties-provider';
 import { RuleInformationProvider } from 'electron/platform/android/rule-information-provider';
 import { RuleInformationProviderType } from 'electron/platform/android/rule-information-provider-type';
-import { AndroidScanResults } from 'electron/platform/android/scan-results';
+import { AndroidScanResults } from 'electron/platform/android/android-scan-results';
 import {
     convertScanResultsToPlatformData,
     ConvertScanResultsToPlatformDataDelegate,
