+++ src/tests/unit/tests/electron/platform/android/rule-information-provider.test.ts
@@ -4,7 +4,7 @@
 import { UnifiedFormattableResolution } from 'common/types/store-data/unified-data-interface';
 import { RuleInformation } from 'electron/platform/android/rule-information';
 import { RuleInformationProvider } from 'electron/platform/android/rule-information-provider';
-import { RuleResultsData } from 'electron/platform/android/scan-results';
+import { RuleResultsData } from 'electron/platform/android/android-scan-results';
 import { buildRuleResultObject } from './scan-results-helpers';
 
 describe('RuleInformationProvider', () => {
